refactor(renderer): split SettingsModal into per-tab components and per-concern stores
SettingsModal was 1489 lines, of which 186-1485 was a single function holding
six independent tabs. Every signal, memo, effect and cleanup timer shared one
scope, so the mobile-connect countdown interval sat three lines from the
updates tab's state. Changing the mobile-connect expiry copy meant opening
1489 lines of TSX plus 1229 lines of CSS.

The seams were already cut: each Tabs.Content is a boundary, and the
computer-use tab was already a one-line delegation to ComputerUseMacSetup.
This copies that pattern to the other five.

SettingsModal.tsx keeps lifetime, stores/*.ts keep logic, Settings*Tab.tsx
keep markup. Every store is created by SettingsModal and passed to its tab,
because Kobalte unmounts an unselected Tabs.Content when the dialog closes -
state owned inside a tab would reset on reopen, and the device table and
hosted-site list would flash their empty state instead of staying on screen
while they refetch. It is also what the footer save bar needs: it reads the
profile draft while another tab is selected.

modalElement stays a plain `let` passed as a bare identifier. Solid's compiler
does not wrap a bare identifier in a getter, so the Select popover's mount
target is captured at the same moment as before; a signal would have made the
read reactive and relocated the popover into the modal's overflow context.

The CSS split extracts the two self-contained namespaces, .settings-mobile-*
and .hosted-sites-*, each carrying its @media fragments in source order.
Both new sheets are imported after settings-modal.css and before
computer-use.css, which deliberately overrides the shared card and row rules
and has to stay last.

The public surface is unchanged, so SettingsModal.test.tsx, the twelve
stories, WorkspaceOverlays.tsx and lazy-views.ts need no edits - their passing
is the proof this is behaviour-neutral.
