+++ src/renderer/src/features/settings/settings-mobile-connect.css
@@ -1,0 +1,319 @@
+.settings-mobile-connect-code-collapse {
+  display: grid;
+  grid-template-rows: 1fr;
+  opacity: 1;
+  transition:
+    grid-template-rows var(--openbot-duration-overlay) var(--openbot-ease-out),
+    opacity var(--openbot-duration-normal) ease-out;
+}
+
+@starting-style {
+  .settings-mobile-connect-code-collapse {
+    grid-template-rows: 0fr;
+    opacity: 0;
+  }
+}
+
+.settings-mobile-connect-code-collapse[data-collapsing] {
+  grid-template-rows: 0fr;
+  opacity: 0;
+}
+
+.settings-mobile-connect-code-collapse-body {
+  min-height: 0;
+  overflow: hidden;
+}
+
+.settings-mobile-connect-code {
+  min-height: 244px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--openbot-space-6);
+  padding: var(--openbot-space-6);
+}
+
+.settings-mobile-connect-qr-stage {
+  position: relative;
+  flex: 0 0 auto;
+}
+
+.settings-mobile-connect-qr-stage .ui-qr-code {
+  transition:
+    opacity var(--openbot-duration-normal) ease-out,
+    filter var(--openbot-duration-overlay) ease-out,
+    transform var(--openbot-duration-overlay) var(--openbot-ease-out);
+}
+
+.settings-mobile-connect-qr-stage[data-success] .ui-qr-code {
+  opacity: 0.18;
+  filter: blur(2px);
+  transform: scale(0.97);
+}
+
+.settings-mobile-connect-success-mark {
+  position: absolute;
+  inset: 0;
+  display: grid;
+  place-items: center;
+  color: var(--openbot-success);
+  animation: settings-mobile-connect-success-enter var(--openbot-duration-overlay) var(--openbot-ease-out) both;
+}
+
+.settings-mobile-connect-success-mark::before {
+  position: absolute;
+  width: 84px;
+  height: 84px;
+  border: 1px solid var(--openbot-success);
+  border-radius: var(--openbot-radius-pill);
+  background: var(--openbot-success-soft);
+  content: "";
+  animation: settings-mobile-connect-success-ring 600ms var(--openbot-ease-out) both;
+}
+
+.settings-mobile-connect-success-mark > svg {
+  position: relative;
+  width: 58px;
+  height: 58px;
+  padding: var(--openbot-space-2);
+  border-radius: var(--openbot-radius-pill);
+  background: var(--openbot-bg-surface-raised);
+  filter: drop-shadow(0 6px 16px var(--openbot-shadow-popover-key));
+  stroke-width: 1.75;
+  animation: settings-mobile-connect-success-pop 420ms var(--openbot-ease-out) both;
+}
+
+.settings-mobile-connect-success-mark > svg path {
+  stroke-dasharray: 24;
+  stroke-dashoffset: 24;
+  animation: settings-mobile-connect-check-draw 360ms var(--openbot-duration-fast) var(--openbot-ease-out) forwards;
+}
+
+.settings-mobile-connect-code-copy {
+  max-width: 240px;
+  display: grid;
+  gap: var(--openbot-space-2);
+}
+
+.settings-mobile-connect-code-title {
+  font-weight: var(--openbot-weight-semibold);
+}
+
+.settings-mobile-connect-expiry {
+  color: var(--openbot-settings-accent-text);
+  font-variant-numeric: tabular-nums;
+}
+
+.settings-mobile-connect-success-title {
+  color: var(--openbot-success);
+}
+
+.settings-mobile-connect-expired {
+  min-height: 196px;
+  display: grid;
+  place-items: center;
+  align-content: center;
+  gap: var(--openbot-space-2);
+  color: var(--openbot-text-secondary);
+  text-align: center;
+}
+
+.settings-mobile-connect-expired > svg {
+  width: calc(var(--openbot-icon-lg) + var(--openbot-space-3));
+  height: calc(var(--openbot-icon-lg) + var(--openbot-space-3));
+  margin-bottom: var(--openbot-space-2);
+}
+
+.settings-mobile-devices {
+  display: grid;
+  gap: var(--openbot-space-3);
+}
+
+.settings-mobile-devices-heading {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--openbot-space-3);
+}
+
+.settings-mobile-devices-heading h3 {
+  margin: 0;
+  font-weight: var(--openbot-weight-medium);
+}
+
+.settings-mobile-devices-states {
+  display: grid;
+}
+
+.settings-mobile-devices-state {
+  display: grid;
+  grid-template-rows: 0fr;
+  opacity: 0;
+  transform: translateY(-4px);
+  transition:
+    grid-template-rows var(--openbot-duration-overlay) var(--openbot-ease-out),
+    opacity var(--openbot-duration-normal) ease-out,
+    transform var(--openbot-duration-overlay) var(--openbot-ease-out);
+}
+
+.settings-mobile-devices-state[data-expanded] {
+  grid-template-rows: 1fr;
+  opacity: 1;
+  transform: translateY(0);
+}
+
+.settings-mobile-devices-state-body {
+  min-height: 0;
+  overflow: hidden;
+}
+
+.settings-mobile-devices-empty {
+  min-height: 88px;
+  display: flex;
+  align-items: center;
+  gap: var(--openbot-space-3);
+  border: 1px solid var(--openbot-border);
+  border-radius: var(--openbot-radius-lg);
+  background: var(--openbot-bg-surface);
+  padding: var(--openbot-space-4);
+  color: var(--openbot-text-muted);
+}
+
+.settings-mobile-devices-empty > svg {
+  width: var(--openbot-icon-lg);
+  height: var(--openbot-icon-lg);
+  flex: 0 0 auto;
+}
+
+.settings-mobile-devices-empty > div {
+  display: grid;
+  gap: var(--openbot-space-1);
+}
+
+.settings-mobile-devices-empty-title {
+  color: var(--openbot-text-secondary);
+  font-weight: var(--openbot-weight-medium);
+}
+
+.settings-mobile-devices-table-frame {
+  overflow-x: auto;
+  border: 1px solid var(--openbot-border);
+  border-radius: var(--openbot-radius-lg);
+  background: var(--openbot-bg-surface);
+}
+
+.settings-mobile-devices-table {
+  width: 100%;
+  min-width: 600px;
+  border-collapse: collapse;
+  color: var(--openbot-text-secondary);
+  font-size: var(--openbot-text-sm);
+  line-height: var(--openbot-leading-sm);
+}
+
+.settings-mobile-devices-table th,
+.settings-mobile-devices-table td {
+  padding: var(--openbot-space-3) var(--openbot-space-4);
+  border-bottom: 1px solid var(--openbot-border);
+  text-align: left;
+  white-space: nowrap;
+}
+
+.settings-mobile-devices-table th {
+  color: var(--openbot-text-muted);
+  font-weight: var(--openbot-weight-medium);
+}
+
+.settings-mobile-devices-table tbody tr:last-child td {
+  border-bottom: 0;
+}
+
+.settings-mobile-device-name {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--openbot-space-2);
+  color: var(--openbot-text-primary);
+  font-weight: var(--openbot-weight-medium);
+}
+
+.settings-mobile-device-name svg {
+  width: var(--openbot-icon-md);
+  height: var(--openbot-icon-md);
+  color: var(--openbot-text-muted);
+}
+
+.settings-mobile-devices-table td.settings-mobile-device-action {
+  width: 1%;
+  text-align: right;
+}
+
+@keyframes settings-mobile-connect-success-enter {
+  from {
+    opacity: 0;
+  }
+}
+
+@keyframes settings-mobile-connect-success-ring {
+  from {
+    opacity: 0.8;
+    transform: scale(0.72);
+  }
+
+  to {
+    opacity: 0;
+    transform: scale(1.35);
+  }
+}
+
+@keyframes settings-mobile-connect-success-pop {
+  from {
+    opacity: 0;
+    transform: scale(0.62) rotate(-8deg);
+  }
+
+  70% {
+    transform: scale(1.08) rotate(1deg);
+  }
+
+  to {
+    opacity: 1;
+    transform: scale(1) rotate(0);
+  }
+}
+
+@keyframes settings-mobile-connect-check-draw {
+  to {
+    stroke-dashoffset: 0;
+  }
+}
+
+@media (max-width: 720px) {
+  .settings-mobile-connect-code {
+    align-items: center;
+    flex-direction: column;
+    text-align: center;
+  }
+
+  .settings-mobile-connect-code-copy {
+    max-width: 320px;
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .settings-mobile-connect-code-collapse,
+  .settings-mobile-connect-qr-stage .ui-qr-code,
+  .settings-mobile-devices-state {
+    transition: none;
+  }
+
+  .settings-mobile-connect-success-mark,
+  .settings-mobile-connect-success-mark::before,
+  .settings-mobile-connect-success-mark > svg,
+  .settings-mobile-connect-success-mark > svg path {
+    animation: none;
+  }
+
+  .settings-mobile-connect-success-mark > svg path {
+    stroke-dashoffset: 0;
+  }
+}
